+++ NEWS.md
@@ -1,6 +1,7 @@
 # excursions (development version)
 
 * Add support for INLA results computed with the `experimental` option.
+* Avoid deprecated Matrix (>=1.4-2) class coercion methods
 
 # excursions 2.5.2